Refactor pipelines fetch environments
