+++ frontend/src/modules/Pipelines/components/PipelineEnvironmentCreateForm.js
@@ -17,23 +17,23 @@ import {
   TextField
 } from '@mui/material';
 import PropTypes from 'prop-types';
-import React, { useEffect, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { Defaults } from 'design';
 import { SET_ERROR, useDispatch } from 'globalErrors';
-import { listEnvironmentGroups, useClient } from 'services';
+import {
+  listEnvironmentGroups,
+  listValidEnvironments,
+  useClient
+} from 'services';
 import { createDataPipelineEnvironment } from '../services';
 
 export const PipelineEnvironmentCreateForm = (props) => {
-  const {
-    environmentOptions,
-    triggerEnvSubmit,
-    pipelineUri,
-    handleCountEnvironmentValid
-  } = props;
+  const { triggerEnvSubmit, pipelineUri, handleCountEnvironmentValid } = props;
   const dispatch = useDispatch();
   const client = useClient();
   const [kvEnvs, setKeyValueEnvs] = useState([]);
   const [mapGroups, setMapGroups] = useState(new Map());
+  const [environmentOptions, setEnvironmentOptions] = useState([]);
   const stageOps = [
     { value: 'dev', label: 'dev' },
     { value: 'test', label: 'test' },
@@ -41,21 +41,29 @@ export const PipelineEnvironmentCreateForm = (props) => {
     { value: 'prod', label: 'prod' },
     { value: 'other', label: 'other' }
   ];
-
-  const environmentOps =
-    environmentOptions && environmentOptions.length > 0
-      ? environmentOptions
-      : [
-          { environmentUri: 'someUri', label: 'some' },
-          { environmentUri: 'someUri', label: 'some2' }
-        ];
+  const fetchEnvironments = useCallback(async () => {
+    const response = await client.query(
+      listValidEnvironments({ filter: Defaults.selectListFilter })
+    );
+    if (!response.errors) {
+      setEnvironmentOptions(
+        response.data.listValidEnvironments.nodes.map((e) => ({
+          ...e,
+          value: e.environmentUri,
+          label: e.label
+        }))
+      );
+    } else {
+      dispatch({ type: SET_ERROR, error: response.errors[0].message });
+    }
+  }, [client, dispatch]);
 
   const fetchGroups = async (environment) => {
     try {
       const response = await client.query(
         listEnvironmentGroups({
           filter: Defaults.selectListFilter,
-          environmentUri: environment.environmentUri
+          environmentUri: environment.value
         })
       );
 
@@ -63,7 +71,7 @@ export const PipelineEnvironmentCreateForm = (props) => {
         setMapGroups(
           new Map(
             mapGroups.set(
-              environment.environmentUri,
+              environment.value,
               response.data.listEnvironmentGroups.nodes
             )
           )
@@ -75,6 +83,14 @@ export const PipelineEnvironmentCreateForm = (props) => {
       dispatch({ type: SET_ERROR, error: e.message });
     }
   };
+
+  useEffect(() => {
+    if (client) {
+      fetchEnvironments().catch((e) =>
+        dispatch({ type: SET_ERROR, error: e.message })
+      );
+    }
+  }, [client, dispatch, fetchEnvironments]);
 
   const handleAddEnvRow = () => {
     if (kvEnvs.length <= 40) {
@@ -101,7 +117,7 @@ export const PipelineEnvironmentCreateForm = (props) => {
         rows[idx].stage = value;
       } else if (field === 'env') {
         rows[idx].environmentLabel = value.label;
-        rows[idx].environmentUri = value.environmentUri;
+        rows[idx].environmentUri = value.value;
       } else {
         rows[idx].samlGroupName = value;
       }
@@ -225,9 +241,9 @@ export const PipelineEnvironmentCreateForm = (props) => {
                                 select
                                 variant="outlined"
                               >
-                                {environmentOps.map((environment) => (
+                                {environmentOptions.map((environment) => (
                                   <MenuItem
-                                    key={environment.environmentUri}
+                                    key={environment.value}
                                     value={environment}
                                   >
                                     {environment.label}
@@ -285,8 +301,8 @@ export const PipelineEnvironmentCreateForm = (props) => {
     </>
   );
 };
+
 PipelineEnvironmentCreateForm.propTypes = {
-  environmentOptions: PropTypes.array.isRequired,
   triggerEnvSubmit: PropTypes.bool.isRequired,
   pipelineUri: PropTypes.string.isRequired,
   handleCountEnvironmentValid: PropTypes.func.isRequired
