+++ frontend/src/modules/Pipelines/views/PipelineCreateForm.js
@@ -6,7 +6,6 @@ import {
   Card,
   CardContent,
   CardHeader,
-  CircularProgress,
   Container,
   FormHelperText,
   Grid,
@@ -17,7 +16,7 @@ import {
 } from '@mui/material';
 import { Formik } from 'formik';
 import { useSnackbar } from 'notistack';
-import { useCallback, useEffect, useState } from 'react';
+import { useState } from 'react';
 import { Helmet } from 'react-helmet-async';
 import { Link as RouterLink, useNavigate } from 'react-router-dom';
 import * as Yup from 'yup';
@@ -25,17 +24,13 @@ import {
   ArrowLeftIcon,
   ChevronRightIcon,
   ChipInput,
-  Defaults,
   useSettings
 } from 'design';
 import { SET_ERROR, useDispatch } from 'globalErrors';
-import {
-  listEnvironmentGroups,
-  listValidEnvironments,
-  useClient
-} from 'services';
+import { useClient } from 'services';
 import { createDataPipeline } from '../services';
 import { PipelineEnvironmentCreateForm } from '../components';
+import { EnvironmentTeamDropdown } from 'modules/Shared';
 
 const PipelineCreateForm = (props) => {
   const navigate = useNavigate();
@@ -43,9 +38,6 @@ const PipelineCreateForm = (props) => {
   const dispatch = useDispatch();
   const client = useClient();
   const { settings } = useSettings();
-  const [loading, setLoading] = useState(true);
-  const [groupOptions, setGroupOptions] = useState([]);
-  const [environmentOptions, setEnvironmentOptions] = useState([]);
   const devOptions = [
     { value: 'cdk-trunk', label: 'CDK Pipelines - Trunk-based' },
     { value: 'trunk', label: 'CodePipeline - Trunk-based' },
@@ -58,56 +50,6 @@ const PipelineCreateForm = (props) => {
   const handleCountEnvironmentValid = (state) => {
     setCountEnvironmentsValid(state);
   };
-
-  const fetchEnvironments = useCallback(async () => {
-    setLoading(true);
-    const response = await client.query(
-      listValidEnvironments({ filter: Defaults.selectListFilter })
-    );
-    if (!response.errors) {
-      setEnvironmentOptions(
-        response.data.listValidEnvironments.nodes.map((e) => ({
-          ...e,
-          value: e.environmentUri,
-          label: e.label
-        }))
-      );
-    } else {
-      dispatch({ type: SET_ERROR, error: response.errors[0].message });
-    }
-    setLoading(false);
-  }, [client, dispatch]);
-
-  const fetchGroups = async (environmentUri) => {
-    try {
-      const response = await client.query(
-        listEnvironmentGroups({
-          filter: Defaults.selectListFilter,
-          environmentUri
-        })
-      );
-      if (!response.errors) {
-        setGroupOptions(
-          response.data.listEnvironmentGroups.nodes.map((g) => ({
-            value: g.groupUri,
-            label: g.groupUri
-          }))
-        );
-      } else {
-        dispatch({ type: SET_ERROR, error: response.errors[0].message });
-      }
-    } catch (e) {
-      dispatch({ type: SET_ERROR, error: e.message });
-    }
-  };
-
-  useEffect(() => {
-    if (client) {
-      fetchEnvironments().catch((e) =>
-        dispatch({ type: SET_ERROR, error: e.message })
-      );
-    }
-  }, [client, dispatch, fetchEnvironments]);
 
   async function submit(values, setStatus, setSubmitting, setErrors) {
     if (!countEnvironmentsValid) {
@@ -123,7 +65,7 @@ const PipelineCreateForm = (props) => {
               label: values.label,
               environmentUri: values.environment.environmentUri,
               description: values.description,
-              SamlGroupName: values.SamlGroupName,
+              SamlAdminGroupName: values.SamlAdminGroupName,
               tags: values.tags,
               devStrategy: values.devStrategy
             }
@@ -157,10 +99,6 @@ const PipelineCreateForm = (props) => {
         dispatch({ type: SET_ERROR, error: err.message });
       }
     }
-  }
-
-  if (loading) {
-    return <CircularProgress />;
   }
 
   return (
@@ -229,7 +167,7 @@ const PipelineCreateForm = (props) => {
               initialValues={{
                 label: '',
                 description: '',
-                SamlGroupName: '',
+                SamlAdminGroupName: '',
                 environment: '',
                 tags: [],
                 devStrategy: 'cdk-trunk'
@@ -239,7 +177,7 @@ const PipelineCreateForm = (props) => {
                   .max(255)
                   .required('*Pipeline name is required'),
                 description: Yup.string().max(5000),
-                SamlGroupName: Yup.string().max(255),
+                SamlAdminGroupName: Yup.string().max(255),
                 environment: Yup.object(),
                 devStrategy: Yup.string().required(
                   '*A CICD strategy is required'
@@ -330,96 +268,13 @@ const PipelineCreateForm = (props) => {
                     <Grid item lg={5} md={6} xs={12}>
                       <Card sx={{ mb: 3 }}>
                         <CardHeader title="CICD" />
-                        <CardContent>
-                          <TextField
-                            fullWidth
-                            error={Boolean(
-                              touched.environment && errors.environment
-                            )}
-                            helperText={
-                              touched.environment && errors.environment
-                            }
-                            label="CICD Environment"
-                            name="environment"
-                            onChange={(event) => {
-                              setFieldValue('SamlGroupName', '');
-                              fetchGroups(
-                                event.target.value.environmentUri
-                              ).catch((e) =>
-                                dispatch({ type: SET_ERROR, error: e.message })
-                              );
-                              setFieldValue('environment', event.target.value);
-                            }}
-                            select
-                            value={values.environment}
-                            variant="outlined"
-                          >
-                            {environmentOptions.map((environment) => (
-                              <MenuItem
-                                key={environment.environmentUri}
-                                value={environment}
-                              >
-                                {environment.label}
-                              </MenuItem>
-                            ))}
-                          </TextField>
-                        </CardContent>
-                        <CardContent>
-                          <TextField
-                            fullWidth
-                            error={Boolean(
-                              touched.SamlGroupName && errors.SamlGroupName
-                            )}
-                            helperText={
-                              touched.SamlGroupName && errors.SamlGroupName
-                            }
-                            label="Team"
-                            name="SamlGroupName"
-                            onChange={(event) => {
-                              setFieldValue(
-                                'SamlGroupName',
-                                event.target.value
-                              );
-                            }}
-                            select
-                            value={values.SamlGroupName}
-                            variant="outlined"
-                          >
-                            {groupOptions.map((group) => (
-                              <MenuItem key={group.value} value={group.value}>
-                                {group.label}
-                              </MenuItem>
-                            ))}
-                          </TextField>
-                        </CardContent>
-                        <CardContent>
-                          <TextField
-                            disabled
-                            fullWidth
-                            label="Region"
-                            name="region"
-                            value={
-                              values.environment
-                                ? values.environment.region
-                                : ''
-                            }
-                            variant="outlined"
-                          />
-                        </CardContent>
-                        <CardContent>
-                          <TextField
-                            disabled
-                            fullWidth
-                            label="Organization"
-                            name="organization"
-                            value={
-                              values.environment
-                                ? values.environment.organization.label
-                                : ''
-                            }
-                            variant="outlined"
-                          />
-                        </CardContent>
+                        <EnvironmentTeamDropdown
+                          setFieldValue={setFieldValue}
+                          handleChange={handleChange}
+                          values={values}
+                          touched={touched}
+                          errors={errors}
+                        />
                         <CardContent>
                           <TextField
                             fullWidth
@@ -448,7 +303,6 @@ const PipelineCreateForm = (props) => {
                     <Grid item lg={12} md={6} xs={12}>
                       <Box sx={{ mt: 3 }}>
                         <PipelineEnvironmentCreateForm
-                          environmentOptions={environmentOptions}
                           triggerEnvSubmit={triggerEnvSubmit}
                           pipelineUri={pipelineUri}
                           handleCountEnvironmentValid={
